soep-core-preprocessing-pipeline/requirements.txt
-i https://pypi.org/simple
gitdb2==2.0.3
gitpython==2.1.10
jinja2==2.10
lxml==4.2.3
markupsafe==1.0
numpy==1.14.5; python_version != '3.3.*'
pandas==0.23.3
python-dateutil==2.7.3
pytz==2018.5
pyyaml==3.13
scipy==1.1.0
six==1.11.0
smmap2==2.0.3
